+++ coefficient_list.php
@@ -153,15 +153,20 @@ $permissiontoread = $user->rights->funding->coefficient->read;
 $permissiontoadd = $user->rights->funding->coefficient->write;
 $permissiontodelete = $user->rights->funding->coefficient->delete;
 
-// Security check
-if (empty($conf->funding->enabled)) accessforbidden('Module not enabled');
-$socid = 0;
-if ($user->socid > 0) {	// Protection if external user
-	//$socid = $user->socid;
+// Security check - Protection if external user
+if (isset($user->socid) && $user->socid > 0) {
+	$action = '';
+	$socid = $user->socid;
+}
+if ($user->socid > 0) accessforbidden();
+//if ($user->socid > 0) $socid = $user->socid;
+//$isdraft = (($object->statut == $object::STATUS_DISABLE) ? 1 : 0);
+//$result = restrictedArea($user, 'funding', $object->id, '', '', 'fk_soc', 'rowid', $isdraft);
+
+if (!isModEnabled("funding")) {
 	accessforbidden();
 }
-//$result = restrictedArea($user, 'funding', $id, '');
-//if (!$permissiontoread) accessforbidden();
+if (!$permissiontoread) accessforbidden();