+++ funding_document.php
@@ -80,14 +80,23 @@ include DOL_DOCUMENT_ROOT.'/core/actions_fetchobject.inc.php'; // Must be includ
 //if ($id > 0 || ! empty($ref)) $upload_dir = $conf->funding->multidir_output[$object->entity?$object->entity:$conf->entity] . "/funding/" . dol_sanitizeFileName($object->id);
 if ($id > 0 || !empty($ref)) $upload_dir = $conf->funding->multidir_output[$object->entity ? $object->entity : $conf->entity]."/funding/".dol_sanitizeFileName($object->ref);
 
+$permissiontoread = $user->rights->funding->read;
+$permissiontoadd = $user->rights->funding->write; // Used by the include of actions_addupdatedelete.inc.php
+
 // Security check - Protection if external user
+if (isset($user->socid) && $user->socid > 0) {
+	$action = '';
+	$socid = $user->socid;
+}
 if ($user->socid > 0) accessforbidden();
 //if ($user->socid > 0) $socid = $user->socid;
-//$result = restrictedArea($user, 'funding', $object->id);
+//$isdraft = (($object->statut == $object::STATUS_DISABLE) ? 1 : 0);
+//$result = restrictedArea($user, 'funding', $object->id, '', '', 'fk_soc', 'rowid', $isdraft);
 
-$permissiontoadd = $user->rights->funding->funding->write; // Used by the include of actions_addupdatedelete.inc.php
-
-
+if (!isModEnabled("funding")) {
+	accessforbidden();
+}
+if (!$permissiontoread) accessforbidden();
 
 /*
  * Actions